requirements.txt
python-telegram-bot==20.7
anthropic==0.20.0
openai==1.14.0
gspread==6.0.2
google-auth==2.28.0
google-api-python-client==2.120.0
aiohttp==3.9.3
python-dotenv==1.0.1
babel==2.14.0
matplotlib==3.8.3
reportlab==4.1.0
